requirements.txt
# Core dependencies
streamlit==1.31.0
langchain==0.1.0
langchain_community==0.0.13
langchain_text_splitters==0.0.1
pinecone-client==3.0.0
python-dotenv==1.0.0
sentence-transformers==2.2.2
transformers==4.36.2
torch==2.1.2
bs4==0.0.1
requests==2.31.0
PyPDF2==3.0.1
docx2txt==0.8
tqdm==4.66.1

# Optional: for better performance with GPU
# accelerate==0.25.0